unleash use temp dir for cache
